RewardEvaluator type must now derive from Number only

+++ bundles/org.palladiosimulator.simexp.core/src/org/palladiosimulator/simexp/core/reward/SimpleRewardSignal.java
@@ -2,13 +2,13 @@ package org.palladiosimulator.simexp.core.reward;
 
 import org.palladiosimulator.simexp.markovian.model.markovmodel.markoventity.impl.RewardImpl;
 
-public class SimpleRewardSignal extends RewardImpl<Double> {
+public class SimpleRewardSignal extends RewardImpl<Integer> {
 
     private final static int POSITIV_REWARD_SIGNAL = 1;
     private final static int NEGATIV_REWARD_SIGNAL = -1;
 
     private SimpleRewardSignal(int value) {
-        super.setValue((double) value);
+        super.setValue(value);
     }
 
     public static SimpleRewardSignal createPositivReward() {
@@ -21,7 +21,7 @@ public class SimpleRewardSignal extends RewardImpl<Double> {
 
     @Override
     public String toString() {
-        return Double.toString(getValue());
+        return Integer.toString(getValue());
     }
 
 }
